+++ CHANGELOG.md
@@ -12,6 +12,8 @@ hi
 - Flipped the Hanging Tendril's item texture vertically to be more consistent with Vanilla.
 - Removed the Firefly's "Nectar" easter egg.
 - Fireflies now once again move inside Display Lanterns that are hanging.
+- Crabs should no longer have an incorrect rotation while climbing downwards.
+- Renamed the `wilderwild:crab_can_hide` tag to `wilderwild:crab_hideable.`
 - Significantly optimized Sculk Sensor rendering with the `Minecraft Live Tendrils` resource pack.
 - Fixed many recipe unlocks being broken in 1.21.2+.
 - Fixed Potted Mycelium Growth not being in the `minecraft:flower_pots` tag.